+++ src/main/java/cart/controller/PageController.java
@@ -1,7 +1,7 @@
-package cart.ui;
+package cart.controller;
 
-import cart.application.ProductService;
 import cart.dao.MemberDao;
+import cart.service.ProductService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
