+++ src/main/java/cart/service/ProductService.java
@@ -1,7 +1,7 @@
-package cart.application;
+package cart.service;
 
-import cart.domain.Product;
 import cart.dao.ProductDao;
+import cart.domain.Product;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
 import org.springframework.stereotype.Service;
